[player] fix seek bugs
add seekto api to seek to an definitely pos
pass accurate time(us) to demuxer for seek

+++ dtdemux/demuxer/demuxer_ts.c
@@ -887,7 +887,7 @@ static int ts_read_frame(demuxer_wrapper_t *wrapper, dt_av_frame_t *frame)
     return DTERROR_READ_EOF;
 }
 
-static int ts_seek_frame(demuxer_wrapper_t *wrapper, int timestamp)
+static int ts_seek_frame(demuxer_wrapper_t *wrapper, int64_t timestamp)
 {
     return 0;
 }
